requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10  
aiosignal==1.3.1  
asgiref==3.8.1    
attrs==24.2.0     
black==24.10.0    
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Django==5.1.3
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
drf-spectacular==0.27.2
frozenlist==1.5.0
generics==7.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
propcache==0.2.0
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
yarl==1.17.1
python-dotenv==1.0.1
psycopg2-binary==2.9.10
